+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 * Fix a bug where an absolute Windows path would be considered an `input:output`
   pair.
 
+* Forbid custom properties that have no values, like `--foo:;`, since they're
+  forbidden by the CSS spec.
+
 ## 1.5.0
 
 * Fix a bug where an importer would be passed an incorrectly-resolved URL when